LIB-64: Add git hook integration for clang-format
This requires that the user runs install.sh to install the hooks, as git hooks are not committed to the repository and have to be installed. The hooks are installed as a copy and have to be re-installed every time until a solution is found that works well on all platforms and not just linux.

The pre-commit hook will guarantee that clang-format is run before committing, removing the need to manually run the hooks on the source files. This requires clang-format to be found on the path.