LIB-48: Support for AppVeyor Continuous Integration
Allows using AppVeyor for continuous integration, which will help determining the state of a commit and PR in various configurations and compilers at the same time. The integration relies on a few pre-built binaries which are only used to reduce the build times drastically.

The main build process is done using Node.js, which comes pre-installed on the AppVeyor machines, and it allows us to do many of the steps in parallel instead of waiting for the other to complete. This effectively uses the resources available on the AppVeyor machines, reducing our total build time slightly.