cmake: Support for proper exporting via CPack and CMake
Adds support for INSTALL and PACKAGE target with proper configurations for including and find_package. libcaffeine can now be included using a very simple find_package command that looks for cmake/<BITS>/CaffeineConfig.cmake, and if it finds it includes that file.

Binary names have been adjusted to be different for each target, to support the import system better. RelWithDebInfo has the original name, Release is suffixed with r, MinSizeRel is suffixed with s, and Debug is suffixed with d. This way all four configurations can be linked to without having to build a different package for each configuration.