Address several possible race conditions.
Use broadcast transition state for state transitions that depend on a previous state.
When failing to transition to live set the stage to not live.
Only end a broadcast from the failed callback when it matches the instance's current broadcast.
Stop a broadcast synchronously to prevent weird issues from starting a broadcast when another is stopping.